refactor: 프로젝트 생성 시의 입력 컴포넌트와 메시지 컴포넌트를 별도의 컴포넌트로 분리 (CM-106)
